V-2.46.18 - CAD Ellipse - option CW (clockwise) added - CATALOG-parts can have subdirectories - Joint can import plane for positioning submodel - bugfix display of reversed Circle - bugfix select position gives wrong parameter for reversed curves - bugfix polygonal representation of trimmed-polygon in contour - bugfix loft-surf with reversed curve - bugfix insert points for "S Polygon < Points" (window GroupEdit) - modification of trimmed curve for boolean surface operations - functions Display/ObjNames ON and Display/ObjDirection ON rewritten

+++ src/ut/ubssur.h
@@ -29,8 +29,8 @@ int USBS_DegElevBspSur (SurBSpl*, Memspc*, SurBSpl*, int, int);
 int USBS_DerivBspSur (SurBSpl*, Memspc*, SurBSpl*, int);
 int USBS_DifKntsKntVecBspSur (int*, double**, Memspc*, int, double*,
 	                      SurBSpl*, int);
-int USBS_DirIndSdBspSur (double**, double**, long*, Point**, Memspc*,
-	                 double**, double**, long*, Point**, Memspc*,
+int USBS_DirIndSdBspSur (double**, double**, int*, Point**, Memspc*,
+	                 double**, double**, int*, Point**, Memspc*,
 		         int, double*, double*, int, int, Point*, int, 
 		         double, int);
 int USBS_EvPtBspSur (Point*, SurBSpl*, double, double, Memspc*);
